caddyhttp: Add optional trailing slash syntax for path matchers
This is an idea taken from the WICG URLPattern draft, that I thought would be a perfect fit for Caddy.

https://github.com/WICG/urlpattern/blob/main/mdn-drafts/QUICK-REFERENCE.md

It's reasonably common that users want to match both `/foo` and `/foo/*` at the same time, but not `/foobar`. This is relevant when some service is being proxied to under a subpath, but you want to make sure there's no "leakage" in the path matching, matching more than will actually function properly.

The URLPattern draft has the `/*?` syntax which means "make the wildcard and slash optional". This is a pretty neat pattern that can fold `/foo` and `/foo/*` as above into just `/foo/*?` to match the exact same thing in one rule.

This means it can also be used for inline path matchers to do both those things at once, usually saving a line of Caddyfile config.

Before:
```
example.com

@service path /service /service/*
reverse_proxy @service localhost:8000
```

After:
```
example.com

reverse_proxy /service/*? localhost:8000
```