chore: Attempt to build with CGO, to verify that we have no CGO deps
Run the build with CGO enabled, to verify that the build still passes, even without installing anything extra.

This test should fail if we add any dependencies that optionally uses CGO.

This is an idea I had that could let us know if some change to our dependencies starts optionally using CGO.

The reason this is useful is because Go actually defaults to enabling CGO if the GOOS/GOARCH build target is the same as the current platform, and CGO is supported on that platform. This means some users were Go builds of Caddy implicitly enabling CGO which would fail due to us previously having an indirect dependency on DataDog/zstd. That dependency is now gone since #4307, so we should no longer have any build issues _when enabling CGO_ if we don't have any dependencies which may optionally use CGO.