logging: Fix `cookie` filter
Fix #4941

The implementation wasn't correct. Headers in logs are actually of type `caddyhttp.LoggableStringArray`, but the implementation was trying to read it from `in.String` which only works if the logged field is actually a string; it's not, so we need to use `in.Interface` and type assert it.

The test was also incorrect for the same reason. I think the tests were written with the assumption that the logs actually work a certain way, but they don't in reality.

Tested with the following Caddyfile, with this command `curl -v -H'Cookie: myvalue=1; sessionid=ST1LLh3r3' http://localhost:8883`

```
{
    debug
    admin off
    servers {
        log_credentials
    }
}

:8883 {
    log {
        format filter {
            wrap json
            fields {
                request>headers>Cookie cookie {
                    replace sessionid REDACTED
                }
            }
        }
    }
}
```