Update adminapi.go
The caddy.Context.AppIsConfigured() method returns true if the named app is present in the config. This is useful to know if an app has been explicitly configured, because calling App() will instantiate an app module even if it's not configured.

However, if the primary use case is to get an app instance only if it has been configured, then maybe we should actually have a method, AppIfConfigured() instead. One-letter change.

This method would return nil if the app is not configured, so if you need the bool value, there you have it; otherwise it returns the configured app module. It would be less clunky than this: