caddyauth: improve handling of multiple providers
Improve the authentication handler in case multiple authentication
providers are set. Currently if multiple providers are present and
at least one of them fails to authenticate it could be that response
contained a mix of multiple providers (and potentially the next
handler).

This commit changes the behavior in the following way:
In case one of the providers succeeds all information of the
providers is discarded and the next handler is called. In case all
providers fail to authenticate the user, the response of any
provider that redirects is written. In case no provider redirects the
response of any provider is written.

Currently there is no way to require all authentication providers to
successfully authenticate, but this behavior wasn't possible in the
previous implementation either.

Fixes #5190