reverseproxy: refactor HTTP transport layer
The earlier implementation smuggles the upstream dial info through context. It made it difficult to accommodate utilizing the standard env vars `HTTP_PROXY`, `HTTPS_PROXY`, and `NO_PROXY` for upstream connections. It also hurt connection caching. The earlier hack was to accommodate the case of upstreams being Unix sockets. The alternative approach is to register a custom protocol with a custom Transport, which is provided by the package github.com/peterbourgon/unixtransport, and let it manage the case of unix sockets.