Fixed an option that was not implemented on stderr
Hello, first of all I love Caddy! I hope this helps.

THE PROBLEM:
In current "83b2697" caddy version is not working the option "capture_stderr" of the directive "php_fastcgi".

HOW TO TEST THE PROBLEM:
If you want to test it, just create a php file with syntax error of any type and pass it over php_fastcgi.

WHEN STARTED THE PROBLEM:
The problem started in the transition between the "f2a7e7c" commit, and the current "83b2697".

WHY STARTED THE PROBLEM:
On 2022 Sept 2, it was launched the commit "83b2697" which changed some files on fastcgi.go.
These changes was titled: fastcgi: Optimize FastCGI transport (#4978).
It was changed a lot of internal functions and they were replaced with a more distributed logic into separate files (which is good).
On those changes was removed a conditional that evaluates "t.CaptureStderr" (line 164, f2a7e7c).
This conditional set the correct logger to be passed to the client (fcgiBackend in f2a7e7c) instance.
So the logger was always set to "noopLogger" (a zap.NewNop()) and no matters you set capture_stderr, it was never registered into log.

HOW TECHNICALLY OCCURRED THE PROBLEM:
Those changes on the "t.CaptureStderr" (line 164, f2a7e7c) behaviour, probably were done because it was assumed that the evaluation of "c.stderr" (line 244, "83b2697") on client.go (another change introduced on "83b2697") would have the same behaviour, and it is right in some way.
I mean, if "c.stderr" is set, it will work as expected.
The problem is that nowhere it was passed the value evaluated on "t.CaptureStderr" to the client instance created, and so, it was never evaluated on "c.stderr" (line 244, "83b2697").

THE SOLUTION:
I must to say that, although it has taken me quite some time to find the solution, it is actually quite simple.
For fixing it I just pass the value in the creation of the client instance, ash show in this pull request.


I hope this little enhancement could help a bit to us.