httpcaddyfile: Fix redir <to> html
Before this commit, the response's Content-Type header was set to
text/plain, which causes some browsers to treat it as plaintext instead
of html, and thereby just showing the html code instead of parsing it
and redirecting as expected.

This commit fixes that issue by setting Content-Type to text/html
Now the browser parses the response body as HTML, which leads to the
redirect working as expected.